Move some functionality to buildSrc
